lang/rust: patch to build with LLVM 16
LLVM 15 and later treat -Wimplicit-function-declaration as errors by
default when using the C99 standard or later [1].

lzma-sys from Rust tries to use gettimeofday(2) unavailable in C99
instead of correctly detecting that FreeBSD provides clock_gettime(2).

Patch lzma-sys to indicate that it can use clock_gettime(2) on FreeBSD
that is defined in sys/time.h and available in C99.

[1] https://reviews.llvm.org/D122983