graphics/mesa-dri: always enable EGL
Don't disable EGL when building Mesa for the hybrid ABI. In such case,
no DRI drivers are compiled, including those we're interested in, e.g.
the Panfrost driver.

Also, Mesa enables EGL by default when some DRI driver is required so we
don't need to enable it with a flag.