www/qt5-webkit: mark for deletion
Qt upstream switched to QtWebEngine quite a while ago, and the maintenance
of QtWebKit was carried out by a separate person.
Unfortunately, no release or updates have been pushed in the recent
years.

Note: there is a fork [1] of the used fork [2], that might be viable for people that
really still need to rely on Qt5Webkit.

[1] https://github.com/movableink/webkit/
[2] https://github.com/qtwebkit/qtwebkit