multimedia/lebiniou*: Take Maintainer'ship
Following the sad news of Olivier's passing, the original author of this
FreeBSD port, I commit to continue maintaining this port in his memory.

The software remains in widespread use, and as a co-author of this port
who assisted Olivier in its FreeBSD porting, I want to ensure that his
work lives on.

Rest in peace, Olivier - oliv3 - Girondel.