editors/openoffice-4: Unmark BROKEN
The undefined symbol ERR_GET_FUNC error was caused by libserf-1.so
expecting to find this symbol in libssl.so, but it was no longer
defined in OpenSSL > 3.0.  This problem was fixed in www/serv 1.3.9_7.
The unsupported option -print-multi-os-directory error comes from
an OpenOffice module configure script and does not break the build.

MFH:		2023Q3