www/chromium: fix linker flags for aarch64
When compiling Chromium with LLVM 16 for aarch64, we hit the same issue
as upstream FreeBSD ports did for i386: Chromium cannot link with ffmpeg
due to text relocations in assembly code.

A similar issue was reported and fixed in the past in upstream Chromium
[1]. Note that the embedded ffmpeg automatically adds -Wl,-z,notext for
lld when Chromium is compiled for amd64 but not arm64.

[1] https://issuetracker.google.com/issues/187786235?pli=1