deskutils/xfce4-notifyd: Properly handle libcanberra dependency
xfce4-notifyd configure script automatically adds a dependency on
libcanberra-gtk3 if found on the system.

To have control on this I added a new CANBERRA option (disabled by
default) to the port to control configure behavior and ensure that
the port does not link to it if the option is disabled.

PR:		273805
MFH:		2023Q3
(cherry picked from commit 4018c2022be41e2e634d970655cfe78c4b997dc0)