devel/binutils: Disable randomly discovered libelf dependency
Users who build binutils from ports source will discover an implied
dependency for elfutils should elfutils be installed at the time
make is run.

PORTREVISION is not bumped because this bug does not affect building of
binary packages. The binary package never has an elfutils dependency
because elfutils is never installed in the poudriere jail prior to
building binutils. This avoids a gratuitous rebuild of binary packages
when none is needed.

PR:	273798