textproc/hs-pandoc: Enable the 'lua' option
This option was accidentally disabled during the update to 3.1.8
due to some glitch in the Haskel support in the ports framework.

The 'lua' option is ON by default in the project. However, Lua
dependencies are only added by 'make make-use-cabal' when
CABAL_FLAGS=lua is present, despite this option being ON by default.

PR:		273729
Reported by:	arrowd@