net/libdnet: Use libpfctl
Several old ioctl commands have been removed in 15/main, which broke
compilation for this port. Rely on libpfctl instead.

There are some breaking pf API changes in 14 which involves regarding
the ioctl syscalls. This port is an old outdated port which explicitly
used those older syscalls. This port also is dependency of net/scapy
which is run in our CI for testing. So fix it with blanket approval.

For more information see:
https://lists.freebsd.org/archives/freebsd-pf/2023-April/000345.html

PR:		273899
Approved by:	portmgr (just-fix-it)
Sponsored by:   Rubicon Communications, LLC ("Netgate")