devel/py-cadquery-pywrap: Switch to llvm11
The port had a runtime dependency on llvm10 however this port also has a
runtime dependency on py-clang which is as of now version 11 that means
that it is based on llvm11/clang11. So in effort to remove llvm10 switch
to llvm11. In case this breaks runtime upstream has more recent versions
which supports llvm13.

PR:		273964
Reported by:	brooks
Approved by:	portmgr (blanket)