x11-wm/stumpwm: Update to 22.11-88 and rework port
Generate a bundled stumpwm binary.

In ca5de48040, the stumpwm port was unbundled so that it had runtime
dependencies on lang/sbcl, a few common lisp ports, as well as the
associated ports with compiled lisp code.  While this made the stumpwm
port lean, it required more maintenance and made it fragile.  Whenever
any of the dependencies changed, stumwpwm would fail to run until it was
rebuilt.  Since that time, sbcl has gained core compression support so
that the bundled stumpwm binary is reduced from about 60 MB to under 12
MB.

Sponsored by:	The FreeBSD Foundation