net/gorss: update to version 0.4, don't force infocmp(1) installation
LookupTerminfo() first tries to find definitions based on the TERM
environment variable and falls back to parsing infocmp(1) output
only if that fails.  The latter is really a method of last resort,
as the performance will be slow.  Do not force this dependency since
it is rarely needed, and most likely due to incorrectly configured
environment which should be fixed instead.

Improve user-friendliness of the program on the first run by copying
installed example configuration and default theme files into user's
home directory (exact location depends on XDG configuration).

PR:	272768