devel/gamin: fix build with lld 17
Building devel/gamin with lld 17 results in the following link error:

  cc -shared  .libs/gam_api.o .libs/gam_data.o .libs/gam_fork.o .libs/gam_error.o .libs/gam_event.o  -lpthread -L/usr/local/lib  -fstack-protector-strong -Wl,--version-script=./gamin_sym.version -fstack-protector-strong -Wl,-soname -Wl,libfam.so.0 -o .libs/libfam.so.0.0.0
  ld: error: version script assignment of 'global' to symbol 'FAMDebugLevel' failed: symbol not defined
  ld: error: version script assignment of 'global' to symbol 'FAMDebug' failed: symbol not defined
  cc: error: linker command failed with exit code 1 (use -v to see invocation)

This is because the FAMDebugLevel and FAMDebug symbols are only defined
when gamin is configured with --enable-debug, which we have no option
for in the port.

So fix it by removing the undefined symbols from the linker version
script.

PR:		274089
MFH:		2023Q4