astro/osmium-tool: Update to 1.16.0
1. Removes libbz2.so from LIB_DEPENDS

I initially thought this was necessary, but saw a warning that it
probably was not necessary. It looks like no other ports had a similar
LIB_DEPENDS, and the lib in question is in /usr/lib (should be in base).

2. Cleans up LZ4 CMake options

In the original submission for this port (ports/269616), a committer noted
that the CMake LZ4 options could be simplified, so I have applied the
suggestions in this patch.

PR:		274117