x11/xfce4-clipman-plugin: Update to 1.6.5
Explicitly disable wayland support for the tiem being.

Rationale:

XFCE wayland support is WIP upstream [1] and not ready for use.
Only some components have been ported and there is no timeframe for
official support at present.

In this port I'm disabling it explicitly because, if wayland libraries
are found, they are linked in and, due to the experimental nature
of the feature, I'd rather not add it as an OPTION at present and
without proper testing of the whole suite.