graphics/darktable: Build explicitly with GCC 12.x on FreeBSD 12.x
On FreeBSD 12.x, it looks like GCC 13+ is incompatible with the version
of libcxx. In the case of darktable, this leads to the following compile
time error:

    In file included from /usr/include/c++/v1/__functional/weak_result_type.h:16,
                     from /usr/include/c++/v1/__functional/invoke.h:14,
                     from /usr/include/c++/v1/__functional_base:15,
                     from /usr/include/c++/v1/string:520,
                     from .../work/darktable-4.4.2/src/external/rawspeed/src/librawspeed/common/ChecksumFile.h:23,
                     from .../work/darktable-4.4.2/src/external/rawspeed/src/librawspeed/common/ChecksumFile.cpp:21:
    /usr/include/c++/v1/type_traits:1770:8: error: expected identifier before '__is_convertible'
     1770 | struct __is_convertible
          |        ^~~~~~~~~~~~~~~~

I have no idea if it is possible to fix or work around this problem in
darktable itself. Thus, on FreeBSD 12.x, we pin the version of GCC to
12. This way, once GCC 13+ bescomes the default version, darktable can
still be built on FreeBSD 12.x, even though the user may have to install
two versions of GCC.

For FreeBSD 13.x and onward, the version of GCC remains unpinned in the
port and the default one will be used.

This solution was discussed with salvadore@ in the following exp-run
request:
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=273397

PR:		273681