{emulators,sysutils}/xen-{kernel,tools}: update to 4.17.2
Also take the opportunity to switch to fetching the sources from the GitHub
mirror based on commit hash.  This makes it easier to update the port when
security fixes or relevant bugfixes are backported to the stable branches,
instead of manually having to cherry pick the changes and add them as patches.

Long term we should look into building emulators/qemu with Xen support so that
the xen-tools package doesn't need to carry its own version of QEMU.

Approved by: bapt (implicit)