www/gitlab-ce: upgrade to 16.4.1
At first make sure you finished the upgrade to 16.3.x!

While running the database migration step, an error can happen.
So maybe test this at first in a test instance and/or have a backup ready.
You should not need it, but for this update it is better to have it ;)

A fix is described in the upgrade manual:
https://gitlab.fechner.net/mfechner/Gitlab-docu/-/blob/master/update/16.3-16.4-freebsd.md#database-migration-fails
The problem itself is described here:
https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8227

But as everyone has a different name for the foreign key that causes this problem
a corrected migration step does maybe not fix the problem for all users.
This is the reason I described steps in the upgrade manual to fix it for all users
and not wait for a upstream migration fix that does maybe not fix it, like this one:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/132686

This version and all further version will not be made available in quartely branch 2023Q4
to give the users the possibilty to first upgrade to required 16.3 version which is
in quarterly branch 2023Q4 available.

Changelog:
https://about.gitlab.com/releases/2023/09/22/gitlab-16-4-released/
https://about.gitlab.com/releases/2023/09/28/security-release-gitlab-16-4-1-released/