devel/py-pyleri: Add py-pyleri 1.4.2
Pyleri is an easy-to-use parser created for SiriDB. We first used lrparsing and
wrote jsleri for auto-completion and suggestions in our web console. Later we
found small issues within the lrparsing module and also had difficulties keeping
the language the same in all projects. That is when we decided to create Pyleri
which can export a created grammar to JavaScript, C, Python, Go and Java.