security/clamav: fix build with lld 17
Building security/clamav with lld 17 results in the following link
errors:

  ld: error: version script assignment of 'CLAMAV_PRIVATE' to symbol 'cli_cvdunpack' failed: symbol not defined
  ld: error: version script assignment of 'CLAMAV_PRIVATE' to symbol 'cli_dbgmsg_internal' failed: symbol not defined
  ld: error: version script assignment of 'CLAMAV_PRIVATE' to symbol 'init_domainlist' failed: symbol not defined
  ld: error: version script assignment of 'CLAMAV_PRIVATE' to symbol 'init_whitelist' failed: symbol not defined
  ld: error: version script assignment of 'CLAMAV_PRIVATE' to symbol 'cli_parse_add' failed: symbol not defined
  ld: error: version script assignment of 'CLAMAV_PRIVATE' to symbol 'cli_bytecode_context_clear' failed: symbol not defined
  cc: error: linker command failed with exit code 1 (use -v to see invocation)

Most of the mentioned symbols no longer exist anywhere in clamav's
source code, except 'cli_cvdunpack', but it is a static function so it
is not visible to external consumers. Remove all the undefined symbols
from the linker version script to fix the build.

PR:		274332
Approved by:	yasu (maintainer)
MFH:		2023Q4