x11/xvattr: resurrect port
Find the distfile online and bring the port back from the attic.
Give it a good dusting off and set LICENSE.

xvattr lets you list the available attributes associated with the Xv
extension to X11. It also allows you to change the values of the
attributes. This can be used to change brightness and so on for programs
that use Xv overlays.

Requested by:	rsjw on libera.chat/#freebsd-ports