x11/fireflies: Deprecate, unmaintained upstream
Upstream is unmaintained for 18 years, needs patches to build, the
bundled libgfx doesn't build at all and the port is unmaintained as
well, therefore deprecate it. Also update WWW/MASTER_SITES to a location
that still exists, just in case someone wants to maintain the port.

Also deprecate graphics/libgfx, it's unmaintained for just as long and
also needs patches to build and x11/fireflies is the last remaining
consumer.