devel/py-dbt-extractor: new port
A tool that processes the most common jinja value templates in dbt model
files. The tool depends on tree-sitter and the tree-sitter-jinja2 library.

The current strategy is for this processor to be 100% certain when it can
accurately extract values from a given model file. Anything less than 100%
certainty returns an exception so that the model can be rendered with python
Jinja instead.

PR:		274578
Author:		Pat Maddox <pat@patmaddox.com>