databases/py-snowflake-connector-python: new port
The Snowflake Connector for Python conforms to the Python DB API 2.0
specification.

The Snowflake Connector for Python provides an interface for developing Python
applications that can connect to Snowflake and perform all standard
operations. It provides a programming alternative to developing applications in
Java or C/C++ using the Snowflake JDBC or ODBC drivers.

The connector has no dependencies on JDBC or ODBC. It can be installed using pip
on Linux, Mac OSX, and Windows platforms where Python 3.8.0 (or higher) is
installed.

PR:		274576
Author:    Pat Maddox <pat@patmaddox.com>