lang/s7: Scheme interpreter
s7 is a Scheme interpreter intended as an extension language for other
applications.  Although it is a descendant of tinyScheme, s7 is closest
as a Scheme dialect to Guile 1.8. I believe it is compatible with r5rs
and r7rs.  It has continuations, ratios, complex numbers, macros,
keywords, hash-tables, multiprecision arithmetic, generalized set!,
unicode, and so on. It does not have syntax-rules or any of its friends,
and it thinks there is no such thing as an inexact integer.

WWW: https://ccrma.stanford.edu/software/snd/snd/s7.html