UPDATING: Fix old portmaster example
The entry dated 20181213 suggests

    portmaster -f `pkg shlib -qR libperl.so.5.26`

to force an update or re-installation of all ports that directly
depend on the given library. This command leads to excessive and
unnecessary port builds, though, since the -f option is passed
down to the up-to-date checks of all dependencies, causing them
to be rebuilt, too.

The correct command is:

    portmaster `pkg shlib -qR libperl.so.5.26`

Reported by:	rkoberman@gmail.com (Kevin Oberman)