devel/libqsbr: fix build on 32 bit platforms
The fix is to make qsbr_epoch_t a 64 bit integer in any case.  The fix
is carefully designed not to change the API on platforms where it is so
already.  Unfortunately this fix requires FreeBSD 13 or newer due to the
old LLVM version included in FreeBSD 12.

While we are at it, hook up test target.

See also:	https://github.com/rmind/libqsbr/issues/12
Approved by:	portmgr (build fix blanket)
MFH:		2023Q4