print/texlive-texmf: Temporarily reduce DEPENDS
This is a large port and due to the dependencies often triggered for
rebuilds. There might be cases where python and texlive-base might not
be necessary hence reduce the temporarily remove the dependencies and
check the outcome. On my cases I did not find any problem as the
runtimes are installed by other necessary packages for those who use it
extensively for writing. If there is a runtime breakage scream at me
please at the earliest.

Remove an unnecesary VARIABLE from previous commit which is for
MAINTAINER_MODE only.

Not bumping as this does not feel trivial and no need to trigger a
rebuild of other ports which depends on doxygen.