science/qt6-quick3dphysics: Add SIMD option
This fixes building on i386 when SSE2 instructions are not available [1].

Vector intrinsics require either SSE2 or NEON intruction sets, so this
option is only available for aarch64, amd64, armv7, and i386. Off by
default on i386, since the package builders build for the least common
denominator and don't assume that the processor has SSE2 like the
headers in PhysX do.

Reported by:	pkg-fallout [1]