net-mgmt/librenms: unbreak lnms
lnms insists upon lnms file is owned by root, it should be owned and ran
by a non-privileged user

Make it chown www because that's what the librenms service runs as.

I'm discussing a solution with upstream as to how to better fix this in
their code. The current solution allows the app to modify its own code.