games/moonlight-qt: Fix build error for 12.4-RELEASE
There is openssl.pc in /usr/libdata/pkgconf on FreeBSD 13.2-RELEASE.
But there is nothing on 12.4-RELEASE.So 'PKGCONF += openssl' in .pro
file will fail on it.  Openssl 1.1.1q is the native part of the system.
It should check for the existence of this dependency without limiting
the openssl version.  I have deleted PKGCONF += openssl and manual
add appropriate LDFLAGS on 12.4-RELEASE

PR:		274791