Mk/Uses/fortran.mk: remove -B from LDFLAGS
That flag makes the compiler use /usr/local/bin/ld from devel/binutils.
The base system was stuck with an old GPLv2 version of binutils at the
time and lang/gcc required a newer version.

With LLVM ld in base it seems to be safe to remove that flag.

Discussed with:	tijl