editors/emacs: Remove USES=makeinfo
The port has USES=makeinfo which helps to build the info files but the
distfile ships with pre built info files in ${WRKSRC}/info hence
makeinfo is not a hard requirement for the building of emacs itself. And
the build time error can easily be avoided with BINARY_ALIAS to
makeinfo.

As we are utilizing the default info files instead of the generated info
hence bumping PORTREVISION.

PR:		274570
Approved by:	emacs (jrm)