www/dooble: Fix build with upcoming Qt 6.6.0
This port builds with -Werror, and due to QSqlDataBase::exec() being
finally marked as deprecated in Qt 6.6.0 [1], it fails to build due to
deprecation warnings. Add upstreamed patch to fix this issue. [2] Port
builds perfectly fine with previous versions of Qt, including 5.x, as it
took 12 years for the soft deprecation in the API notes [3] to actually
take effect. Upstream notes this was an automatic-completion error. [2]

While here: clean up USE_QT for both the qt5 and qt6 flavors and
remove checks for long unsupported (3+ years) versions of Qt 5.x. [4]

[1] https://doc.qt.io/qt-6/qsqldatabase-obsolete.html
[2] https://github.com/textbrowser/dooble/pull/199
[3] https://doc.qt.io/qt-5/qsqldatabase.html

Approved by:	portmgr (blanket)
With hat:	kde@ [4]