archivers/snappy-java: remove run_depends on openjdk8
This port is a library used by other Java ports. It does not run
itself so does not need a run_depends on openjdk8.
Solves: unifi7 currently installing openjdk11 and openjdk8 because of this.

Not bumping portrevision as it does not change the functionality of the port.

tested:
poudriere stage-qa: ok

PR: 274902