devel/paq-nvim: New port: Minimal Neovim plugin manager
Paq is a minimal, simple to use packer manager for Neovim written in
Lua and is aimed for users who make use of Lua for their configuration
language. It is written to make use of existing Neovim features,
and to have a simple and familiar interface for declaring packages.

Website: https://github.com/savq/paq-nvim

Differential Revision:	https://reviews.freebsd.org/D42185