www/qt6-webengine: Really disable PCH
Due to the way qt_evaluate_config_expression()[1] parses conditions,
De Morgan's Laws do not apply. This function is internal to the Qt
build process and this behavior may be by design for simplicity.
NOT (A OR B) defies logic and always evaluates to NOT A OR B, so we
must express this as NOT A AND NOT B instead to get the desired result.

Disabling PCH (pre-compiled headers) greatly decreases build time,
especially with ccache and should fix the build on arm64 (not tested).

../../../../../qtwebengine-everywhere-src-6.5.3/src/3rdparty/chromium/base/strings/string_util.h:51:3:
error: non-const lvalue reference to type '__builtin_va_list' cannot bind to a value of unrelated type 'va_list' (aka 'std::__va_list')
  va_start(arguments, format);
  ^~~~~~~~~~~~~~~~~~~~~~~~~~~
/usr/include/sys/_stdarg.h:45:49: note: expanded from macro 'va_start'
  #define       va_start(ap, last)      __builtin_va_start((ap), (last))
                                                           ^~~~

[1] https://code.qt.io/cgit/qt/qtbase.git/tree/cmake/QtFeature.cmake?h=6.5.3#n84

(cherry picked from commit 089da4a7a764183a8d63d22dedc12537f90b50db)