cpu-microcode-intel: Add knobs for RC RUN_DEPENDS and split files
Split UCODE files as well as dependency on sysutils/cpu-microcode-rc
are only required for use with cpucontrol(8) and update method two
respectively. The first method needs neither so we can spare a few
MiB and a superfluous port.
Given that the automatic update method one is superior, the RC knob
should be deprecated and removed eventually. At that point, split
UCODE files still will be useful for developers and if users want
to update the microcode without a reboot (but otherwise rely on the
loader for automatic updates, i. e. method one).

Reviewed by:	jrm (w/ style suggestions)
Approved by:	jrm (upcoming maintainer)
Differential Revision:	https://reviews.freebsd.org/D41904