lang/julia: partly remove the dependency on gfortran
Theoretically this port should not depend on gfortran, because we do not
build the bundled OpenBLAS (use the system package) nor SuiteSparse (GPL
libraries are deactivated).

It is certainly possible to totally remove the dependency on gfortran,
but this would require some more intrusive patches: this one only removes
the linkage with the libraries provided by GCC (among them libgomp).

Remark: this does not fix the failure reported at
<https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=273219#c37>.