net/nss_ldap: fix the build on both 13.2 and 14.0+
The original configure script attempted to detect the presence of
Heimdal, but it failed with LLVM 15+ due to a type casting bug.
commit 515552b2d2664dc02c2385519a7fedb855f46632 fixed the bug with
LLVM 15+, but inadvertently broke the build for earlier versions.

This commit corrects the configure syntax, fixing the build on 12.4 and
13.2.

PR:		270465
Approved by:	mikael (ports)
Sponsored by:	Axcient
MFH:		2023Q4