lang/clover: Restrict llvm to max version 16
Clover cannot (yet?) be compiled with LLVM17, it misses
llvm/Transforms/IPO/PassManagerBuilder.h
For now set a max version of llvm to 16 for clover only, people should still
be able to use llvm17 for other mesa componant if they are not interested in clover.

While here clean up a bit the USES=llvm for mesa port, we already set noexport
in mesa-dri/Makefile.common so no need to set it in all the ports.

Sponsored by:	Beckhoff Automation GmbH & Co. KG
PR:	275290