devel/qtcreator: Use llvm17 as recommended
QtCreator would crash in the settings dialog if not rebuilt after
an update of Qt packages to 6.6.1 release. Make build and run depend
on llvm17 as recommended upstream, for better clangd compatibility:

https://www.qt.io/blog/qt-creator-12-c-code-model-update

Since llvm17 seems to be more strict about non-standard alloca stack
allocation, patch that out of the internal SQLite build config. This
matches the config of databases/sqlite3 in ports.

PR:		275636