devel/libnotify: fix build with lld 17 (take 2)
The fix for bug 274108 broke running the notify-send command, resulting
in:

  ld-elf.so.1: /usr/local/lib/libglib-2.0.so.0: Undefined symbol "environ"

This is because libnotify uses the same linker version script for both
its libnotify.so shared library and for its notify-send executable.

The linker version script places all symbols that do not begin with
"notify_" into the local scope, making them effectively hidden. This is
fine for a shared library, but not for an executable, since it will also
hide the symbols "environ" and "__progname", which are required for most
programs.

Fix it by not using the linker version script for the notify-send
executable, so no required symbols are hidden.

PR:		275104, 274108
MFH:		2023Q4