lang/gcc12: install ISO_Fortran_binding.h in a separate directory
ISO_Fortran_binding.h is sometimes needed for the support of Fortran
2018, e.g. MPICH must find it to build mpi_f08.mod.

ATM it is installed in a directory depending on $CONFIGURE_TARGET with
a complex path, and it is not possible to point CFLAGS towards this dir
because other headers (like stdatomic.h) would conflict with those from
the system.

Therefore create a new directory /usr/local/include/gcc12 to install
this header.

PR:		275431
Approved by:	salvadore@ (maintainer)