audio/alsa-utils: aplaymidi doesn't play at the proper tempo
Add temporary patch to fix an issue with tempo events

aplaymidi does not process tempo events because the part
that handles tempo events was not rewritten sufficiently.
It may be executed as a different event in songs that
have tempo changes in the middle of the song.

A similar fix was submitted to upstream as
https://github.com/alsa-project/alsa-utils/issues/241

PR:		275349
Reported by:	Tatsuki Makino <tatsuki_makino@hotmail.com>