www/gitlab-ce: use update patch if not already updated to 16.6.0/1
The patch was updated, so use it if you have not already used the previous one.
This is required as the old version cannot be downloaded anymore.

If you have already upgraded to 16.6.1, this patch can be ignored.
This is also the reason why I do not bump the version.