sysutils/u-boot-bhyve-arm64: Don't build depend on e2fsprogs or gnutls
These are only needed when building mkeficapsule, which is enabled by
default based on EFI_CAPSULE_ON_DISK. For the arm64 bhyve port neither
are set and so we don't actually need the build dependency.

This is true for many of the other ports upstream, too, upstream is just
lazy and doesn't bother to make the dependency conditional (given
nothing even ships the tool probably it should even be turned off
explicitly for every port, with an updated u-boot-tools providing it
when necessary), but I don't want to go and hunt all those down, since
we really don't care about anything other than u-boot-bhyve-arm64.