emulators/qmc2: Sync with upstream; switch to QWebengine
- Sync with upstream
- Switch browser from QWebKit to QWebengine and remove DEPRECATED/
  EXPIRATION_DATE
- Pet port(clippy|lint)(1)
- Simplify COMMENT
- Move to external dependencies archivers/libarchive and
  archivers/minizip-ng
- Drop desktop-file-utils from USES -- the port does not register
  any MIME Types
- Move some post-patch target work to static patching
- Remove hard-coded binary and manpage names in do-install target
- Add patches to better integrate with system-wide configuration
- Bump minimum runtime requirement on MAME to 0.259

In memoriam of René Reucher, author of qmc and respected MAME community
member: May his labour in bringing nostalgia and making perhaps the best
era of arcade entertainment accessible to the masses be forever inscribed
upon this digital stele.

On behalf of the FreeBSD Project, we greatly appreciate his contributions
to the FOSS community and offer our belated condolences to his family
and friends.

PR:		275324