kde.mk: add virtual category 'kde-devel'
Previously we assumed that a port of KDE Gear 24.02-beta1 will use Qt6.
So kde:6 was used to determine that the applications version is beta1.
However, not all applications will be ported in time of the release as
it seems, so we need to have a better way to select the version.

For now, the KDE Gear-ports can use (for beta1)

	CATEGORIES=$main kde kde-applications kde-devel

resepectively (for stable version)

	CATEGORIES=$main kde kde-applications

to get the proper DISTVERSION value and environment set up.