www/py-django50: New port
Django is a high-level Python Web framework that encourages rapid
development and clean, pragmatic design.

Developed and used over the past two years by a fast-moving online-news
operation, Django was designed from scratch to handle two challenges:
the intensive deadlines of a newsroom and the stringent requirements of
experienced Web developers. It has convenient niceties for developing
content-management systems, but it's an excellent tool for building any
Web site.

Currently this is marked to IGNORE with default python version 3.9. If
someone needs to use this they have to build it with:
DEFAULT_VERSIONS=python=3.10

This is mainly the work from wen@ but was reverted as it broke the bulk
build.