net/socat: Add configurable IPv4/IPv6 preference option
In socat version 1.8.0.0, a new feature was introduced allowing users to
configure the default IP version preference at build time using the
`--enable-default-ipv` option. This change resulted in a default
preference for IPv6, which led to unexpected behavior for some users.

To address this, a new option has been added to the FreeBSD port of
socat. By default, the port restores the previous behavior where IPv4
is preferred. However, users now have the option to explicitly specify a
preference for IPv6 if desired.

This change aims to align with the Principle of Least Astonishment
(POLA) and provide users with flexibility in choosing their preferred IP
version.

PR:		275653 (based on)