lang/gnat12: bootstrap from previous assets by default
All assets have been built, and it is now possible to bootstrap without
lang/gcc6-aux!

But the option GCC6AUX it not removed, just in case, if for some reason
we had to rebuild everything from the beginning.

Do not bump PORTREVISION: the resulting package should not be impacted.